Update IDA Pro integration script
