--- r2ida/.travis.yml
@@ -1,5 +1,0 @@
-language: python
-python:
-  - "2.7"
-  - "3.4"
-
