--- r2ida/doc/migration-from-ida.md
@@ -1,4 +1,0 @@
-Migration from IDA to Radare2
-=============================
-
-https://radare.gitbooks.io/radare2book/content/debugger/migration.html
