+++ rz-ida/README.md
@@ -1,0 +1,12 @@
+# rz-ida
+
+This tool facilitates the import of metadata from the IDA Pro to use inside Rizin.
+It supports the parsing of the `*.idb` files and `*.idc` exported scripts.
+
+**Note:** This requires the [python-idb](https://github.com/williballenthin/python-idb) installed
+and available for python imports.
+
+It is recommended to use the latest version from the PyPI:
+```sh
+pip3 install -U --user python-idb
+```
